refactor: replace lock management with lockService and lockRepository for improved synchronization
