+++ web-app/src/app/api/sync/agents/route.ts
@@ -6,7 +6,8 @@ import { getEnvSecrets } from "@/config/env.secrets";
 import { postRegistryEntry } from "@/lib/api/generated/registry";
 import { getRegistryClient } from "@/lib/api/registry-service.client";
 import { authenticateCronSecret } from "@/lib/auth/utils";
-import { acquireLock, prisma, unlockLock } from "@/lib/db/repositories";
+import { lockRepository, prisma } from "@/lib/db/repositories";
+import { lockService } from "@/lib/services";
 import { AgentStatus, Lock, PricingType } from "@/prisma/generated/client";
 
 const LOCK_KEY = "agents-sync";
@@ -21,7 +22,7 @@ async function agentSync() {
   // Start a transaction to ensure atomicity
   let lock: Lock;
   try {
-    lock = await acquireLock(LOCK_KEY, getEnvSecrets().INSTANCE_ID);
+    lock = await lockService.acquireLock(LOCK_KEY, getEnvSecrets().INSTANCE_ID);
   } catch (error) {
     if (error instanceof Error && error.message === "LOCK_IS_LOCKED") {
       return NextResponse.json(
@@ -52,7 +53,7 @@ async function agentSync() {
     } catch (error) {
       console.error("Error in sync operation:", error);
     } finally {
-      const unlocked = await unlockLock(lock.key);
+      const unlocked = await lockRepository.unlockByKey(lock.key);
       if (!unlocked) {
         console.error("Failed to unlock lock");
       }
