+++ web-app/src/lib/services/index.ts
@@ -2,6 +2,7 @@ export * from "./agent";
 export * from "./credit";
 export * from "./invitation";
 export * from "./job";
+export * from "./lock.service";
 export * from "./organization";
 export * from "./stripe";
 export * from "./user.service";
